# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JOB API",
        "description": "API da plataforma JOB para intermediação de serviços.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/negociacoes/{negociacao}/pagamento": {
            "get": {
                "tags": [
                    "Negociações - Pagamentos"
                ],
                "summary": "Exibe o pagamento da negociação",
                "description": "Retorna os dados do pagamento vinculado à negociação.",
                "operationId": "0891f652ebc0318e06e9e5714994ac9e",
                "parameters": [
                    {
                        "name": "negociacao",
                        "in": "path",
                        "description": "ID da negociação",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamento encontrado"
                    },
                    "403": {
                        "description": "Usuário não participa da negociação"
                    },
                    "404": {
                        "description": "Pagamento não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/negociacoes/{negociacao}/pagamento/informar": {
            "post": {
                "tags": [
                    "Negociações - Pagamentos"
                ],
                "summary": "Informa que o pagamento foi realizado",
                "description": "Permite ao contratante informar que realizou o pagamento. O pagamento passa para AGUARDANDO_PROCESSAMENTO e a negociação para PROCESSANDO_PAGAMENTO.",
                "operationId": "97fe5c02a31366dfb0e7cc1c44dbedb3",
                "parameters": [
                    {
                        "name": "negociacao",
                        "in": "path",
                        "description": "ID da negociação",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "observacao": {
                                        "description": "Observação opcional sobre o pagamento.",
                                        "type": "string",
                                        "example": "Pagamento realizado via Pix.",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pagamento informado com sucesso"
                    },
                    "403": {
                        "description": "Somente o contratante pode informar o pagamento"
                    },
                    "422": {
                        "description": "Pagamento não pode ser informado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/negociacoes/{negociacao}/pagamento/confirmar": {
            "post": {
                "tags": [
                    "Negociações - Pagamentos"
                ],
                "summary": "Confirma o pagamento pela plataforma",
                "description": "Confirma, para fins de demonstração, que o pagamento foi identificado e validado. A negociação passa para FECHADA, a publicação passa para ENCERRADO, as demais negociações da mesma publicação são encerradas e um trabalho é criado para o contratado.",
                "operationId": "8959dec15a65a52fa73ff8e208b50931",
                "parameters": [
                    {
                        "name": "negociacao",
                        "in": "path",
                        "description": "ID da negociação",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamento validado, negociação fechada, publicação encerrada e trabalho criado"
                    },
                    "404": {
                        "description": "Pagamento ou status não encontrado"
                    },
                    "422": {
                        "description": "Pagamento não está aguardando processamento"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação do Laravel Sanctum",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}